fix: migrate types to primitives
