+++ bindings/rust-bindings/src/embed.rs
